Show crew for flights

+++ app/src/main/java/com/crewly/roster/RosterHelper.kt
@@ -1,16 +1,16 @@
 package com.crewly.roster
 
 import android.annotation.SuppressLint
-import com.crewly.app.CrewlyDatabase
+import com.crewly.BuildConfig
 import com.crewly.app.RxModule
 import com.crewly.aws.AwsRepository
 import com.crewly.duty.Airport
 import com.crewly.duty.Flight
 import com.crewly.duty.Sector
 import com.crewly.logging.LoggingManager
+import com.crewly.repositories.CrewRepository
 import io.reactivex.Completable
 import io.reactivex.Scheduler
-import io.reactivex.functions.BiFunction
 import org.joda.time.DateTime
 import javax.inject.Inject
 import javax.inject.Named
@@ -21,8 +21,8 @@ import javax.inject.Singleton
  */
 @Singleton
 class RosterHelper @Inject constructor(
-  private val crewlyDatabase: CrewlyDatabase,
   private val awsRepository: AwsRepository,
+  private val crewRepository: CrewRepository,
   private val rosterRepository: RosterRepository,
   private val loggingManager: LoggingManager,
   @Named(RxModule.IO_THREAD) private val ioThread: Scheduler
@@ -48,28 +48,56 @@ class RosterHelper @Inject constructor(
     crewCode: String,
     roster: Roster
   ): Completable =
-    clearDatabase()
-      .mergeWith(crewlyDatabase.dutyDao().insertDuties(roster.duties))
-      .mergeWith(crewlyDatabase.sectorDao().insertSectors(roster.sectors))
-      .doOnComplete { updateNetworkFlights(
-        crewCode = crewCode,
-        newSectors = roster.sectors
-      )}
-
-  @SuppressLint("CheckResult")
-  fun updateNetworkFlights(
-    crewCode: String,
-    newSectors: List<Sector>
-  ) {
     awsRepository
       .getFlightsForCrewMember(
         crewCode = crewCode
       )
-      .zipWith(rosterRepository.fetchSectorsBetween(
+      .onErrorReturn { listOf() }
+      .flatMap { flights ->
+        awsRepository
+          .getCrewMembers(
+            userIds = flights.fold(mutableSetOf<Pair<String, Int>>()) { ids, flight ->
+              flight.departureSector.crew.forEach { id ->
+                ids.add(id to flight.departureSector.company.id)
+              }
+              ids
+            }.toList()
+          )
+          .map { crew -> flights to crew }
+      }
+      // Populate roster with crew from network flights
+      .doOnSuccess { (flights, _) ->
+        roster.sectors.forEach { sector ->
+          val networkFlight = flights.find { flight -> flight.departureSector.flightId == sector.flightId }
+          if (networkFlight != null) sector.crew = networkFlight.departureSector.crew
+        }
+      }
+      .flatMapCompletable { (flights, crew) ->
+        rosterRepository.deleteRosterFromToday()
+          .mergeWith(rosterRepository.insertOrReplaceRoster(
+            roster = roster
+          ))
+          .mergeWith(crewRepository.insertOrUpdateCrew(
+            crew = crew
+          ))
+          .doOnComplete { updateNetworkFlights(
+            crewCode = crewCode,
+            newSectors = roster.sectors,
+            flights = flights
+          )}
+      }
+
+  @SuppressLint("CheckResult")
+  fun updateNetworkFlights(
+    crewCode: String,
+    newSectors: List<Sector>,
+    flights: List<Flight>
+  ) {
+    rosterRepository.fetchSectorsBetween(
         crewCode = crewCode,
         startTime = DateTime.now().withTimeAtStartOfDay(),
         endTime = DateTime.now().plusDays(14).withTimeAtStartOfDay()
-      ), BiFunction<List<Flight>, List<Sector>, SectorFetchData> { networkFlights, localSectors ->
+      ).map { localSectors ->
         // Only save the first sector of each day
         val groupedSectors = newSectors
           .sortedBy { it.departureTime.millis }
@@ -77,11 +105,11 @@ class RosterHelper @Inject constructor(
           .flatMap { it.value.take(1) }
 
         SectorFetchData(
-          networkFlights = networkFlights,
+          networkFlights = flights,
           localSectors = localSectors,
           rosterSectors = groupedSectors
         )
-      })
+      }
       .map { (networkFlights, localSectors, rosterSectors) ->
         val sectorsCrewRemovedFrom = localSectors.minus(rosterSectors)
         val sectorsToDelete = sectorsCrewRemovedFrom.filter { sector ->
@@ -99,19 +127,6 @@ class RosterHelper @Inject constructor(
         val sectorsToSave = rosterSectors
           .minus(sectorsToDelete)
           .plus(sectorsToEdit)
-          .map { sector ->
-            // Populate sectors with network flight data
-            networkFlights.find { networkFlight ->
-              val hasNetworkFlight = sector == networkFlight.departureSector
-              if (hasNetworkFlight) {
-                sector.crew = networkFlight.departureSector.crew
-                sector.crew.remove(crewCode)
-              }
-              hasNetworkFlight
-            }
-
-            sector
-          }
 
         SectorRequestData(
           sectorsToDelete = sectorsToDelete,
@@ -154,28 +169,21 @@ class RosterHelper @Inject constructor(
           }
       }
       .flatMapCompletable { (flightsToDelete, flightsToSave) ->
-        awsRepository
-          .deleteFlights(
-            flights = flightsToDelete
-          )
-          .mergeWith(awsRepository.createOrUpdateFlights(
-            flights = flightsToSave
-          ))
+        if (BuildConfig.DEBUG) {
+          Completable.complete()
+        } else {
+          awsRepository
+            .deleteFlights(
+              flights = flightsToDelete
+            )
+            .mergeWith(awsRepository.createOrUpdateFlights(
+              flights = flightsToSave
+            ))
+        }
       }
       .subscribeOn(ioThread)
       .subscribe({}, { error ->
         loggingManager.logError(error)
       })
-  }
-
-  private fun clearDatabase(): Completable {
-    val currentDay = DateTime().withTimeAtStartOfDay()
-
-    return crewlyDatabase.dutyDao()
-      .deleteAllDutiesFrom(currentDay.millis)
-      .mergeWith(
-        crewlyDatabase.sectorDao()
-          .deleteAllSectorsFrom(currentDay.millis)
-      )
   }
 }
